training/requirements.base.txt
diffusers==0.27.0
transformers==4.38.0
accelerate==0.27.0
safetensors==0.4.2
compel==2.0.2
peft==0.10.0
Pillow>=10.0.0
numpy>=1.26.0

